refactor: 💡 extract logger config
