+++ tinyexp/examples/resnet_exp.py
@@ -13,16 +13,10 @@ from torch.optim import SGD
 from torch.optim.lr_scheduler import StepLR
 from torchvision import datasets, transforms
 
-from tinyexp import (
-    CheckpointCfgMixin,
-    RayCfgMixin,
-    RedisCfgMixin,
-    TinyExp,
-    WandbCfgMixin,
-    store_and_run_exp,
-)
+from tinyexp import TinyExp, store_and_run_exp
 from tinyexp.dataset.sampler import InfiniteSampler
 from tinyexp.exceptions import UnknownAcceleratorTypeError
+from tinyexp.exp_mixins import CheckpointCfgMixin, LoggerMixin, RayCfgMixin, RedisCfgMixin, WandbCfgMixin
 
 
 def transform_template_imagenet(
@@ -167,7 +161,7 @@ class RedisCachedImageFolder:
 
 
 @dataclass(repr=False)
-class ResNetExp(TinyExp, RayCfgMixin, RedisCfgMixin, CheckpointCfgMixin, WandbCfgMixin):
+class ResNetExp(TinyExp, RayCfgMixin, RedisCfgMixin, CheckpointCfgMixin, WandbCfgMixin, LoggerMixin):
     mode: str = "train"
     max_train_epochs: int = 90
     max_train_steps: int = -1
