+++ tinyexp/exp_mixins/__init__.py
@@ -1,8 +1,3 @@
-from .mixins import CheckpointCfgMixin, RayCfgMixin, RedisCfgMixin, WandbCfgMixin
+from .basic_mixins import CheckpointCfgMixin, LoggerMixin, RayCfgMixin, RedisCfgMixin, WandbCfgMixin
 
-__all__ = [
-    "CheckpointCfgMixin",
-    "RayCfgMixin",
-    "RedisCfgMixin",
-    "WandbCfgMixin",
-]
+__all__ = ["CheckpointCfgMixin", "LoggerMixin", "RayCfgMixin", "RedisCfgMixin", "WandbCfgMixin"]
